+++ core/vtk/ttkPersistenceDiagram/ttkPersistenceDiagram.cpp
@@ -212,11 +212,23 @@ int ttkPersistenceDiagram::dispatch(
   vtkUnstructuredGrid *outputCTPersistenceDiagram,
   vtkDataArray *const inputScalarsArray,
   const scalarType *const inputScalars,
+  scalarType *outputScalars,
+  SimplexId *outputOffsets,
+  int *outputMonotonyOffsets,
   const SimplexId *const inputOrder,
   const triangulationType *triangulation) {
 
   int status{};
   std::vector<ttk::PersistencePair> CTDiagram{};
+
+  if(BackEnd == BACKEND::APPROXIMATE_TOPOLOGY) {
+    std::cout << "Chosen approx" << std::endl;
+    double *range = inputScalarsArray->GetRange(0);
+    this->setDeltaApproximate(range[1] - range[0]);
+    this->setOutputScalars(outputScalars);
+    this->setOutputOffsets(outputOffsets);
+    this->setOutputMonotonyOffsets(outputMonotonyOffsets);
+  }
 
   status = this->execute(CTDiagram, inputScalars, inputOrder, triangulation);
 
@@ -228,7 +240,7 @@ int ttkPersistenceDiagram::dispatch(
   }
 
   setPersistenceDiagram(outputCTPersistenceDiagram, CTDiagram,
-                        inputScalarsArray, inputScalars, triangulation);
+                        inputScalarsArray, outputScalars, triangulation);
 
   return 1;
 }
@@ -274,12 +286,33 @@ int ttkPersistenceDiagram::RequestData(vtkInformation *ttkNotUsed(request),
   }
 #endif
 
+  vtkNew<ttkSimplexIdTypeArray> outputOffsets{};
+  outputOffsets->SetNumberOfComponents(1);
+  outputOffsets->SetNumberOfTuples(inputScalars->GetNumberOfTuples());
+  outputOffsets->SetName("outputOffsets");
+
+  vtkNew<vtkIntArray> outputMonotonyOffsets{};
+  outputMonotonyOffsets->SetNumberOfComponents(1);
+  outputMonotonyOffsets->SetNumberOfTuples(inputScalars->GetNumberOfTuples());
+  outputMonotonyOffsets->SetName("outputMonotonyffsets");
+  outputMonotonyOffsets->FillComponent(0, 0);
+
+  vtkSmartPointer<vtkDataArray> outputScalars
+    = vtkSmartPointer<vtkDataArray>::Take(inputScalars->NewInstance());
+  outputScalars->SetNumberOfComponents(1);
+  outputScalars->SetNumberOfTuples(inputScalars->GetNumberOfTuples());
+  outputScalars->DeepCopy(inputScalars);
+  outputScalars->SetName("Cropped");
+
   int status{};
   ttkVtkTemplateMacro(
     inputScalars->GetDataType(), triangulation->getType(),
     status = this->dispatch(
       outputCTPersistenceDiagram, inputScalars,
       static_cast<VTK_TT *>(ttkUtils::GetVoidPointer(inputScalars)),
+      static_cast<VTK_TT *>(ttkUtils::GetVoidPointer(outputScalars)),
+      static_cast<SimplexId *>(ttkUtils::GetVoidPointer(outputOffsets)),
+      static_cast<int *>(ttkUtils::GetVoidPointer(outputMonotonyOffsets)),
       static_cast<SimplexId *>(ttkUtils::GetVoidPointer(offsetField)),
       static_cast<TTK_TT *>(triangulation->getData())));